Add some documentation for the FM11NC08S driver
