+++ app/Cargo.toml
@@ -20,7 +20,7 @@ usb-device = "0.2.3"
 usbd-serial = "0.1.0"
 
 # board
-lpc55-hal = { git = "https://github.com/nickray/lpc55-hal", branch = "main" }
+lpc55-hal = { git = "https://github.com/nickray/lpc55-hal", branch = "timer-duration" }
 lpcxpresso55 = { path = "../boards/lpcxpresso55", optional = true }
 prototype-bee = { path = "../boards/prototype-bee", optional = true }
 board-traits = { path = "../boards/board-traits" }
