+++ boards/lpcxpresso55/Cargo.toml
@@ -13,5 +13,5 @@ board-traits = {path = "../board-traits"}
 
 [dependencies.lpc55-hal]
 git = "https://github.com/nickray/lpc55-hal"
-branch = "main"
+branch = "timer-duration"
 features = ["littlefs", "rtic-peripherals"]
