+++ boards/prototype-bee/Cargo.toml
@@ -14,7 +14,7 @@ board-traits = {path = "../board-traits"}
 
 [dependencies.lpc55-hal]
 git = "https://github.com/nickray/lpc55-hal"
-branch = "main"
+branch = "timer-duration"
 features = ["littlefs", "rtic-peripherals"]
 # path = "../../../../lpc55-hal"
