+++ components/fm11nc08/src/lib.rs
@@ -1,36 +1,17 @@
+//! `nfc-device` chip driver implementation for the FM11NC08 NFC Channel Chip
+//!
+//! This chip allows adding NFC functionality, including energy harvesting,
+//! to an existing microcontroller design. This implementation only covers
+//! the ISO 14443-A, Level 4, SPI variant.
+//!
+//! [fm-chip-url]: http://www.fm-chips.com/nfc-channel-ics.html
+//! [fmsh-url]: http://eng.fmsh.com/2ac114a8-ce2d-aa9d-cef4-ad3919eeb513/
+
 #![no_std]
 
-pub mod device;
+pub mod chip;
 
-pub mod traits;
-
-// #[derive(Copy, Clone, Debug, PartialEq)]
-// pub enum State {
-//     NewSession(u8),
-//     Continue(u8),
-// }
-
-// pub enum Error {
-//     NewSession,
-//     NoActivity,
-// }
-
-// pub type Result<T> = core::result::Result<T, Error>;
-
-// pub trait Device {
-//     fn read(&mut self, buf: &mut [u8]) -> Result<State>;
-
-//     fn send(&mut self,buf: &[u8]) -> Result<()>;
-
-//     // typical value: 128
-//     fn frame_size(&self) -> usize;
-// }
-
-
-pub use device::{
-    FM11NC08,
-    Configuration,
-    Register,
-};
+pub use chip::Fm11Nc08S as Chip;
+pub use chip::types::Configuration;
 
 logging::add!(logger);
