+++ components/nfc-device/Cargo.toml
@@ -13,7 +13,6 @@ heapless = "0.5.5"
 
 iso7816 = { path = "../iso7816", features = ["contactless-interchange"] }
 logging = { path = "../logging" }
-fm11nc08 = { path = "../fm11nc08" }
 interchange = { path = "../interchange" }
 
 [features]
