+++ components/nfc-device/src/lib.rs
@@ -1,8 +1,8 @@
 #![no_std]
 
-pub mod types;
+pub mod chip;
+pub mod driver;
 
-pub mod iso14443;
-pub use iso14443::*;
+pub use driver::*;
 
 logging::add!(logger);
