--- components/nfc-device/src/types.rs
@@ -1,1 +1,0 @@
-
